# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/storage/buckets": {
            "get": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Lists all storage buckets owned by the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "List buckets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Bucket"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Creates a new storage bucket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Create a bucket",
                "parameters": [
                    {
                        "description": "Bucket creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Bucket"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Bucket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.Response"
                        }
                    }
                }
            }
        },
        "/storage/buckets/{bucket}": {
            "delete": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Deletes a storage bucket",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Delete a bucket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket name",
                        "name": "bucket",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.Response"
                        }
                    }
                }
            }
        },
...
        "/storage/buckets/{bucket}/versioning": {
            "patch": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Enables or disables object versioning for the specified bucket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Set bucket versioning",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket name",
                        "name": "bucket",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Versioning request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/storage/cluster/status": {
            "get": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Returns the status of all nodes in the distributed storage cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Get storage cluster status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StorageCluster"
                        }
                    }
                }
            }
        },
        "/storage/multipart/:id": {
            "delete": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Cancels the upload and deletes all uploaded parts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Abort multipart upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Upload ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/storage/multipart/:id/complete": {
            "post": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Assembles all uploaded parts into the final object",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Complete multipart upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Upload ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Object"
                        }
                    }
                }
            }
        },
        "/storage/multipart/:id/parts": {
            "put": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Uploads a data chunk (part) for the specified multipart upload",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Upload a part",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Upload ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Part number",
                        "name": "part",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Part"
                        }
                    }
                }
            }
        },
        "/storage/versions/{bucket}/{key}": {
            "get": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Gets a list of all versions of a specific object",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "List object versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket name",
                        "name": "bucket",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Object key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Object"
                            }
                        }
                    }
                }
            }
        },
        "/storage/{bucket}": {
            "get": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Gets a list of all objects within a specific bucket",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "List objects in a bucket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket name",
                        "name": "bucket",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Object"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.Response"
                        }
                    }
                }
            }
        },
...
        "/storage/{bucket}/{key}/multipart": {
            "post": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Creates a new multipart upload session for a bucket and key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Initiate multipart upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket name",
                        "name": "bucket",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Object key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.MultipartUpload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Bucket": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "encryption_enabled": {
                    "type": "boolean"
                },
                "encryption_key_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "versioning_enabled": {
                    "type": "boolean"
                }
            }
        },
...
        "domain.MultipartUpload": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.Object": {
            "type": "object",
            "properties": {
                "arn": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_latest": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "version_id": {
                    "type": "string"
                }
            }
        },
        "domain.Part": {
            "type": "object",
            "properties": {
                "etag": {
                    "type": "string"
                },
                "part_number": {
                    "type": "integer"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "upload_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.StorageCluster": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.StorageNode"
                    }
                }
            }
        },
        "domain.StorageNode": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "host:port",
                    "type": "string"
                },
                "data_dir": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_seen": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "httputil.Meta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "httputil.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "meta": {
                    "$ref": "#/definitions/httputil.Meta"
                }
            }
        },
...
    }
}